Add files via upload
A demo project developed as part of the curriculum - Distributed Databases in under graduation. It is a desktop application that allows the members to register themselves in the Sports Club, the Admin can have an access to the database and view, delete and update the records as per his/her need. It is developed using Microsoft Visual Studio. MySQL is used to handle data and perform all the insert, delete, update and select commands.